feat: Implemented list users API
