+++ server/services/v1/auth/gotrue.go
@@ -49,8 +49,6 @@ const (
 	InvitationStatusExpired  = "EXPIRED"
 )
 
-var InvalidInvitationCodeErr = errors.Unauthenticated("Failed to verify invitation code")
-
 type gotrue struct {
 	AuthConfig config.AuthConfig
 }
@@ -461,223 +459,6 @@ func (g *gotrue) GetAccessToken(ctx context.Context, req *api.GetAccessTokenRequ
 		}, nil
 	}
 	return nil, errors.InvalidArgument("Failed to GetAccessToken: reason = unsupported grant_type, it has to be one of [refresh_token, client_credentials]")
-}
-
-func (g *gotrue) CreateInvitations(ctx context.Context, req *api.CreateInvitationsRequest) (*api.CreateInvitationsResponse, error) {
-	for _, invitation := range req.Invitations {
-		err := createInvitation(ctx, invitation.GetEmail(), invitation.GetRole(), invitation.GetInvitationSentByName(), g)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return &api.CreateInvitationsResponse{}, nil
-}
-
-func createInvitation(ctx context.Context, email string, role string, invitationSentByName string, g *gotrue) error {
-	if email == "" {
-		return errors.InvalidArgument("Email must be specified")
-	}
-	if role == "" {
-		return errors.InvalidArgument("Role must be specified")
-	}
-
-	namespace, err := request.GetNamespace(ctx)
-	if err != nil {
-		log.Err(err).Msg("Failed to get namespace while creating invitation")
-		return errors.Internal("Could not create user invitation")
-	}
-
-	currentSub, err := request.GetCurrentSub(ctx)
-	if err != nil {
-		log.Err(err).Msg("Failed to get current sub while creating invitation")
-		return errors.Internal("Could not create user invitation")
-	}
-
-	expirationTime := time.Now().UnixMilli() + (g.AuthConfig.UserInvitations.ExpireAfterSec * 1000)
-	createInvitationPayload := CreateInvitationPayload{
-		Email:           email,
-		Role:            role,
-		TigrisNamespace: namespace,
-		CreatedBy:       currentSub,
-		CreatedByName:   invitationSentByName,
-		ExpirationTime:  expirationTime,
-	}
-	createInvitationPayloadBytes, err := jsoniter.Marshal(createInvitationPayload)
-	if err != nil {
-		log.Err(err).Msg("Failed to marshal CreateUserInvitationPayload struct to json bytes")
-		return errors.Internal("Could not create user invitation")
-	}
-
-	_, err = invitationsCall(ctx, createInvitationPayloadBytes, "/invitations", http.MethodPost, g)
-	if err != nil {
-		log.Err(err).Msg("Failed to create user invitation")
-		return errors.Internal("Could not create user invitation")
-	}
-	log.Debug().Str("email", email).Int64("expiration_time", expirationTime).Msg("Created user invitation")
-	return nil
-}
-
-func (g *gotrue) DeleteInvitations(ctx context.Context, req *api.DeleteInvitationsRequest) (*api.DeleteInvitationsResponse, error) {
-	if req.GetEmail() == "" {
-		return nil, errors.InvalidArgument("Email must be specified")
-	}
-
-	if req.GetStatus() != "" {
-		err := validateInvitationStatusInput(req.GetStatus())
-		if err != nil {
-			return nil, err
-		}
-	}
-	namespace, err := request.GetNamespace(ctx)
-	if err != nil {
-		log.Err(err).Msg("Failed to get namespace while deleting invitation")
-		return nil, errors.Internal("Could not delete user invitation")
-	}
-
-	currentSub, err := request.GetCurrentSub(ctx)
-	if err != nil {
-		log.Err(err).Msg("Failed to get current sub while deleting invitation")
-		return nil, errors.Internal("Could not delete user invitation")
-	}
-
-	deleteInvitationsPayload := DeleteInvitationsPayload{
-		Email:           req.GetEmail(),
-		CreatedBy:       currentSub,
-		TigrisNamespace: namespace,
-		Status:          req.GetStatus(),
-	}
-	deleteInvitationsPayloadBytes, err := json.Marshal(deleteInvitationsPayload)
-	if err != nil {
-		log.Err(err).Msg("Failed to marshal delete invitation payload to json bytes")
-		return nil, errors.Internal("Could not delete user invitation")
-	}
-
-	_, err = invitationsCall(ctx, deleteInvitationsPayloadBytes, "/invitations", http.MethodDelete, g)
-	if err != nil {
-		log.Err(err).Msg("Failed to create user invitation")
-		return nil, errors.Internal("Could not create user invitation")
-	}
-
-	log.Debug().Str("email", req.GetEmail()).Str("status", req.GetStatus()).Msg("Deleted user invitation(s)")
-	return &api.DeleteInvitationsResponse{}, nil
-}
-
-func (g *gotrue) ListInvitations(ctx context.Context, req *api.ListInvitationsRequest) (*api.ListInvitationsResponse, error) {
-	if req.GetStatus() != "" {
-		err := validateInvitationStatusInput(req.GetStatus())
-		if err != nil {
-			return nil, err
-		}
-	}
-	namespace, err := request.GetNamespace(ctx)
-	if err != nil {
-		log.Err(err).Msg("Failed to get namespace while listing invitations")
-		return nil, errors.Internal("Could not list user invitations")
-	}
-
-	currentSub, err := request.GetCurrentSub(ctx)
-	if err != nil {
-		log.Err(err).Msg("Failed to get current sub while listing invitations")
-		return nil, errors.Internal("Could not list user invitations")
-	}
-	var path string
-	if req.GetStatus() == "" {
-		path = fmt.Sprintf("/invitations?created_by=%s&tigris_namespace=%s", currentSub, namespace)
-	} else {
-		path = fmt.Sprintf("/invitations?created_by=%s&tigris_namespace=%s&status=%s", currentSub, namespace, req.GetStatus())
-	}
-
-	listInvitationsRes, err := invitationsCall(ctx, nil, path, http.MethodGet, g)
-	if err != nil {
-		log.Err(err).Msg("Failed to create list invitations")
-		return nil, errors.Internal("Could not list user invitations")
-	}
-
-	// parse JSON response
-	var invitations []*api.Invitation
-	err = json.Unmarshal(listInvitationsRes, &invitations)
-	if err != nil {
-		log.Err(err).Msg("Failed to deserialize list user invitations response into JSON")
-		return nil, errors.Internal("Could not list user invitations")
-	}
-
-	return &api.ListInvitationsResponse{Invitations: invitations}, nil
-}
-
-func (g *gotrue) VerifyInvitation(ctx context.Context, req *api.VerifyInvitationRequest) (*api.VerifyInvitationResponse, error) {
-	if req.GetEmail() == "" {
-		return nil, errors.InvalidArgument("Email must be specified")
-	}
-	if req.GetCode() == "" {
-		return nil, errors.InvalidArgument("Code must be specified")
-	}
-
-	verifyInvitationPayload := VerifyInvitationPayload{
-		Email: req.GetEmail(),
-		Code:  req.GetCode(),
-	}
-	verifyInvitationPayloadBytes, err := json.Marshal(verifyInvitationPayload)
-	if err != nil {
-		log.Err(err).Msg("Failed to marshal verify invitation payload to json bytes")
-		return nil, errors.Internal("Could not verify user invitation")
-	}
-	verifyInvitationResBytes, err := invitationsCall(ctx, verifyInvitationPayloadBytes, "/invitations/verify", http.MethodPost, g)
-	if err == InvalidInvitationCodeErr {
-		return nil, err
-	}
-	if err != nil {
-		log.Err(err).Msg("Failed to verify invitation")
-		return nil, errors.Internal("Could not verify user invitation")
-	}
-
-	var verifyUserInvitationResponse api.VerifyInvitationResponse
-	err = jsoniter.Unmarshal(verifyInvitationResBytes, &verifyUserInvitationResponse)
-	if err != nil {
-		log.Err(err).Msg("Failed to JSON deserialize gotrue's verify user response")
-		return nil, errors.Internal("Could not verify user invitation")
-	}
-	return &api.VerifyInvitationResponse{
-		TigrisNamespace: verifyUserInvitationResponse.GetTigrisNamespace(),
-		Role:            verifyUserInvitationResponse.GetRole(),
-	}, nil
-}
-
-func invitationsCall(ctx context.Context, payload []byte, path string, method string, g *gotrue) ([]byte, error) {
-	payloadReader := bytes.NewReader(payload)
-
-	client := &http.Client{}
-	invitationReq, err := http.NewRequestWithContext(ctx, method, fmt.Sprintf("%s%s", g.AuthConfig.Gotrue.URL, path), payloadReader)
-	if err != nil {
-		log.Err(err).Msgf("Failed to create invitation request for path: %s", path)
-		return nil, err
-	}
-	invitationReq.Header.Add("X-JWT-AUD", g.AuthConfig.PrimaryAudience)
-	invitationReq.Header.Add("Content-Type", "application/json")
-
-	invitationRes, err := client.Do(invitationReq)
-	if err != nil {
-		log.Err(err).Msgf("Failed to create invitation request for path: %s", path)
-		return nil, err
-	}
-	defer invitationRes.Body.Close()
-
-	if invitationRes.StatusCode != http.StatusOK {
-		if invitationRes.StatusCode == http.StatusUnauthorized {
-			return nil, InvalidInvitationCodeErr
-		} else {
-			log.Error().Int("status", invitationRes.StatusCode).Msgf("Received non OK status from gotrue while performing invitation operation at path: %s", path)
-			return nil, errors.Internal("Received non OK status")
-		}
-	}
-
-	invitationResBody, err := io.ReadAll(invitationRes.Body)
-	if err != nil {
-		log.Err(err).Msgf("Failed to read invitationResBody body for operation at path: %s", path)
-		return nil, errors.Internal("Failed to read invitationResBody body for operation at path: %s", path)
-	}
-	defer invitationRes.Body.Close()
-
-	return invitationResBody, nil
 }
 
 func createUser(ctx context.Context, createUserPayload []byte, aud string, userType string, g *gotrue) error {
